+++ app/system/modules/info/views/info.php
@@ -13,7 +13,7 @@
         </div>
 
     </div>
-    <div class="uk-flex-item-1">
+    <div class="pk-width-content">
 
         <ul id="tab-content" class="uk-switcher uk-margin">
             <li>