+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## WIP
+
+### Fixed
+- Fixed wrapping sidebar if content in main column is to large
+
 ## 0.11.0 (March 30, 2016)
 
 ### Added
@@ -10,7 +15,7 @@
 - Added global params object to view
 - Added file picker
 - Added support for script tag attributes 'defer' and 'async'
- 
+
 ### Changed
 - Transfer widget and menu positions on theme change
 - Image-, video-, link-picker: Preserve existing attributes